requirements.txt
selenium==4.1.3
wget==3.2
python-dotenv==0.17.1
webdriver-manager==3.8.4